GNORM-00: Removed superfluous inheritance classes.
